hex-dma: dma-queue support for alias queues and cached dma
